@@ -1,15 +1,33 @@
-// Code generated by mockery v2.32.4. DO NOT EDIT.
+// Code generated by mockery; DO NOT EDIT.
+// github.com/vektra/mockery
+// template: testify
 
 package crud
 
-import mock "github.com/stretchr/testify/mock"
+import (
+	mock "github.com/stretchr/testify/mock"
+)
+
+// NewMockRepository creates a new instance of MockRepository. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewMockRepository[T any](t interface {
+	mock.TestingT
+	Cleanup(func())
+}) *MockRepository[T] {
+	mock := &MockRepository[T]{}
+	mock.Mock.Test(t)
+
+	t.Cleanup(func() { mock.AssertExpectations(t) })
+
+	return mock
+}
 
 // MockRepository is an autogenerated mock type for the Repository type
-type MockRepository[T interface{}] struct {
+type MockRepository[T any] struct {
 	mock.Mock
 }
 
-type MockRepository_Expecter[T interface{}] struct {
+type MockRepository_Expecter[T any] struct {
 	mock *mock.Mock
 }
 
@@ -17,22 +35,25 @@ func (_m *MockRepository[T]) EXPECT() *MockRepository_Expecter[T] {
 	return &MockRepository_Expecter[T]{mock: &_m.Mock}
 }
 
-// Delete provides a mock function with given fields: id
-func (_m *MockRepository[T]) Delete(id string) error {
-	ret := _m.Called(id)
+// Delete provides a mock function for the type MockRepository
+func (_mock *MockRepository[T]) Delete(id string) error {
+	ret := _mock.Called(id)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Delete")
+	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(string) error); ok {
-		r0 = rf(id)
+	if returnFunc, ok := ret.Get(0).(func(string) error); ok {
+		r0 = returnFunc(id)
 	} else {
 		r0 = ret.Error(0)
 	}
-
 	return r0
 }
 
 // MockRepository_Delete_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'Delete'
-type MockRepository_Delete_Call[T interface{}] struct {
+type MockRepository_Delete_Call[T any] struct {
 	*mock.Call
 }
 
@@ -44,49 +65,57 @@ func (_e *MockRepository_Expecter[T]) Delete(id interface{}) *MockRepository_Del
 
 func (_c *MockRepository_Delete_Call[T]) Run(run func(id string)) *MockRepository_Delete_Call[T] {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(string))
+		var arg0 string
+		if args[0] != nil {
+			arg0 = args[0].(string)
+		}
+		run(
+			arg0,
+		)
 	})
 	return _c
 }
 
-func (_c *MockRepository_Delete_Call[T]) Return(_a0 error) *MockRepository_Delete_Call[T] {
-	_c.Call.Return(_a0)
+func (_c *MockRepository_Delete_Call[T]) Return(err error) *MockRepository_Delete_Call[T] {
+	_c.Call.Return(err)
 	return _c
 }
 
-func (_c *MockRepository_Delete_Call[T]) RunAndReturn(run func(string) error) *MockRepository_Delete_Call[T] {
+func (_c *MockRepository_Delete_Call[T]) RunAndReturn(run func(id string) error) *MockRepository_Delete_Call[T] {
 	_c.Call.Return(run)
 	return _c
 }
 
-// GetAll provides a mock function with given fields:
-func (_m *MockRepository[T]) GetAll() ([]T, error) {
-	ret := _m.Called()
+// GetAll provides a mock function for the type MockRepository
+func (_mock *MockRepository[T]) GetAll() ([]T, error) {
+	ret := _mock.Called()
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetAll")
+	}
 
 	var r0 []T
 	var r1 error
-	if rf, ok := ret.Get(0).(func() ([]T, error)); ok {
-		return rf()
+	if returnFunc, ok := ret.Get(0).(func() ([]T, error)); ok {
+		return returnFunc()
 	}
-	if rf, ok := ret.Get(0).(func() []T); ok {
-		r0 = rf()
+	if returnFunc, ok := ret.Get(0).(func() []T); ok {
+		r0 = returnFunc()
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).([]T)
 		}
 	}
-
-	if rf, ok := ret.Get(1).(func() error); ok {
-		r1 = rf()
+	if returnFunc, ok := ret.Get(1).(func() error); ok {
+		r1 = returnFunc()
 	} else {
 		r1 = ret.Error(1)
 	}
-
 	return r0, r1
 }
 
 // MockRepository_GetAll_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'GetAll'
-type MockRepository_GetAll_Call[T interface{}] struct {
+type MockRepository_GetAll_Call[T any] struct {
 	*mock.Call
 }
 
@@ -102,8 +131,8 @@ func (_c *MockRepository_GetAll_Call[T]) Run(run func()) *MockRepository_GetAll_
 	return _c
 }
 
-func (_c *MockRepository_GetAll_Call[T]) Return(_a0 []T, _a1 error) *MockRepository_GetAll_Call[T] {
-	_c.Call.Return(_a0, _a1)
+func (_c *MockRepository_GetAll_Call[T]) Return(vs []T, err error) *MockRepository_GetAll_Call[T] {
+	_c.Call.Return(vs, err)
 	return _c
 }
 
@@ -112,34 +141,36 @@ func (_c *MockRepository_GetAll_Call[T]) RunAndReturn(run func() ([]T, error)) *
 	return _c
 }
 
-// GetById provides a mock function with given fields: id
-func (_m *MockRepository[T]) GetById(id string) (*T, error) {
-	ret := _m.Called(id)
+// GetById provides a mock function for the type MockRepository
+func (_mock *MockRepository[T]) GetById(id string) (*T, error) {
+	ret := _mock.Called(id)
+
+	if len(ret) == 0 {
+		panic("no return value specified for GetById")
+	}
 
 	var r0 *T
 	var r1 error
-	if rf, ok := ret.Get(0).(func(string) (*T, error)); ok {
-		return rf(id)
+	if returnFunc, ok := ret.Get(0).(func(string) (*T, error)); ok {
+		return returnFunc(id)
 	}
-	if rf, ok := ret.Get(0).(func(string) *T); ok {
-		r0 = rf(id)
+	if returnFunc, ok := ret.Get(0).(func(string) *T); ok {
+		r0 = returnFunc(id)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(*T)
 		}
 	}
-
-	if rf, ok := ret.Get(1).(func(string) error); ok {
-		r1 = rf(id)
+	if returnFunc, ok := ret.Get(1).(func(string) error); ok {
+		r1 = returnFunc(id)
 	} else {
 		r1 = ret.Error(1)
 	}
-
 	return r0, r1
 }
 
 // MockRepository_GetById_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'GetById'
-type MockRepository_GetById_Call[T interface{}] struct {
+type MockRepository_GetById_Call[T any] struct {
 	*mock.Call
 }
 
@@ -151,49 +182,57 @@ func (_e *MockRepository_Expecter[T]) GetById(id interface{}) *MockRepository_Ge
 
 func (_c *MockRepository_GetById_Call[T]) Run(run func(id string)) *MockRepository_GetById_Call[T] {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(string))
+		var arg0 string
+		if args[0] != nil {
+			arg0 = args[0].(string)
+		}
+		run(
+			arg0,
+		)
 	})
 	return _c
 }
 
-func (_c *MockRepository_GetById_Call[T]) Return(_a0 *T, _a1 error) *MockRepository_GetById_Call[T] {
-	_c.Call.Return(_a0, _a1)
+func (_c *MockRepository_GetById_Call[T]) Return(v *T, err error) *MockRepository_GetById_Call[T] {
+	_c.Call.Return(v, err)
 	return _c
 }
 
-func (_c *MockRepository_GetById_Call[T]) RunAndReturn(run func(string) (*T, error)) *MockRepository_GetById_Call[T] {
+func (_c *MockRepository_GetById_Call[T]) RunAndReturn(run func(id string) (*T, error)) *MockRepository_GetById_Call[T] {
 	_c.Call.Return(run)
 	return _c
 }
 
-// Save provides a mock function with given fields: entity
-func (_m *MockRepository[T]) Save(entity *T) (*T, error) {
-	ret := _m.Called(entity)
+// Save provides a mock function for the type MockRepository
+func (_mock *MockRepository[T]) Save(entity *T) (*T, error) {
+	ret := _mock.Called(entity)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Save")
+	}
 
 	var r0 *T
 	var r1 error
-	if rf, ok := ret.Get(0).(func(*T) (*T, error)); ok {
-		return rf(entity)
+	if returnFunc, ok := ret.Get(0).(func(*T) (*T, error)); ok {
+		return returnFunc(entity)
 	}
-	if rf, ok := ret.Get(0).(func(*T) *T); ok {
-		r0 = rf(entity)
+	if returnFunc, ok := ret.Get(0).(func(*T) *T); ok {
+		r0 = returnFunc(entity)
 	} else {
 		if ret.Get(0) != nil {
 			r0 = ret.Get(0).(*T)
 		}
 	}
-
-	if rf, ok := ret.Get(1).(func(*T) error); ok {
-		r1 = rf(entity)
+	if returnFunc, ok := ret.Get(1).(func(*T) error); ok {
+		r1 = returnFunc(entity)
 	} else {
 		r1 = ret.Error(1)
 	}
-
 	return r0, r1
 }
 
 // MockRepository_Save_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'Save'
-type MockRepository_Save_Call[T interface{}] struct {
+type MockRepository_Save_Call[T any] struct {
 	*mock.Call
 }
 
@@ -205,37 +244,46 @@ func (_e *MockRepository_Expecter[T]) Save(entity interface{}) *MockRepository_S
 
 func (_c *MockRepository_Save_Call[T]) Run(run func(entity *T)) *MockRepository_Save_Call[T] {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(*T))
+		var arg0 *T
+		if args[0] != nil {
+			arg0 = args[0].(*T)
+		}
+		run(
+			arg0,
+		)
 	})
 	return _c
 }
 
-func (_c *MockRepository_Save_Call[T]) Return(_a0 *T, _a1 error) *MockRepository_Save_Call[T] {
-	_c.Call.Return(_a0, _a1)
+func (_c *MockRepository_Save_Call[T]) Return(v *T, err error) *MockRepository_Save_Call[T] {
+	_c.Call.Return(v, err)
 	return _c
 }
 
-func (_c *MockRepository_Save_Call[T]) RunAndReturn(run func(*T) (*T, error)) *MockRepository_Save_Call[T] {
+func (_c *MockRepository_Save_Call[T]) RunAndReturn(run func(entity *T) (*T, error)) *MockRepository_Save_Call[T] {
 	_c.Call.Return(run)
 	return _c
 }
 
-// Update provides a mock function with given fields: id, entity
-func (_m *MockRepository[T]) Update(id string, entity *T) error {
-	ret := _m.Called(id, entity)
+// Update provides a mock function for the type MockRepository
+func (_mock *MockRepository[T]) Update(id string, entity *T) error {
+	ret := _mock.Called(id, entity)
+
+	if len(ret) == 0 {
+		panic("no return value specified for Update")
+	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(string, *T) error); ok {
-		r0 = rf(id, entity)
+	if returnFunc, ok := ret.Get(0).(func(string, *T) error); ok {
+		r0 = returnFunc(id, entity)
 	} else {
 		r0 = ret.Error(0)
 	}
-
 	return r0
 }
 
 // MockRepository_Update_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'Update'
-type MockRepository_Update_Call[T interface{}] struct {
+type MockRepository_Update_Call[T any] struct {
 	*mock.Call
 }
 
@@ -248,31 +296,28 @@ func (_e *MockRepository_Expecter[T]) Update(id interface{}, entity interface{})
 
 func (_c *MockRepository_Update_Call[T]) Run(run func(id string, entity *T)) *MockRepository_Update_Call[T] {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(string), args[1].(*T))
+		var arg0 string
+		if args[0] != nil {
+			arg0 = args[0].(string)
+		}
+		var arg1 *T
+		if args[1] != nil {
+			arg1 = args[1].(*T)
+		}
+		run(
+			arg0,
+			arg1,
+		)
 	})
 	return _c
 }
 
-func (_c *MockRepository_Update_Call[T]) Return(_a0 error) *MockRepository_Update_Call[T] {
-	_c.Call.Return(_a0)
+func (_c *MockRepository_Update_Call[T]) Return(err error) *MockRepository_Update_Call[T] {
+	_c.Call.Return(err)
 	return _c
 }
 
-func (_c *MockRepository_Update_Call[T]) RunAndReturn(run func(string, *T) error) *MockRepository_Update_Call[T] {
+func (_c *MockRepository_Update_Call[T]) RunAndReturn(run func(id string, entity *T) error) *MockRepository_Update_Call[T] {
 	_c.Call.Return(run)
 	return _c
-}
-
-// NewMockRepository creates a new instance of MockRepository. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-// The first argument is typically a *testing.T value.
-func NewMockRepository[T interface{}](t interface {
-	mock.TestingT
-	Cleanup(func())
-}) *MockRepository[T] {
-	mock := &MockRepository[T]{}
-	mock.Mock.Test(t)
-
-	t.Cleanup(func() { mock.AssertExpectations(t) })
-
-	return mock
 }
